+++ packages/emp/src/dts/dtsThread.ts
@@ -12,9 +12,11 @@ parentPort.on('message', async (payload: any) => {
     const pathStr = path.join(appAbsSrc, '**/*.(ts|tsx|vue)').replace(/\\/g, '/')
     const dtslist = await glob([pathStr])
     // console.log('dtslist', dtslist, appSrc, `${appSrc}/**/*.(ts|tsx)`)
-    await Promise.all(dtslist.map(async d => {
-       await dts.emit(d)
-    }))
+    await Promise.all(
+      dtslist.map(async d => {
+        await dts.emit(d)
+      }),
+    )
     dts.createFile()
     parentPort.postMessage('finish')
   }