+++ projects/mf-app/src/bootstrap.tsx
@@ -1,4 +1,4 @@
 import ReactDOM from 'react-dom'
 import RouterComp from 'src/RouterComp'
-
+import './normalize.css'
 ReactDOM.render(<RouterComp />, document.getElementById('emp-root'))