+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ## Bugfixes
 + swc 取消 `loose`，解决语法使用的差异问题 如 [link](https://github.com/swc-project/swc/issues/6627),问题清单持续修复 [link](https://github.com/efoxTeam/emp/discussions/327)
 + @efox/plugin-vue-3@1.2.0 修复vue3中使用lang=tsx时的报错问题 by [MAXLZ1](https://github.com/efoxTeam/emp/pull/333)
++ @efox/plugin-vue-2@1.2.2 同时支持 Component 与 src inline [DEMO](https://github.com/efoxTeam/emp/blob/main/projects/vue-2-base/src/views/Home.vue#L15)
 
 # 2.5.x
 > 全面支持 ESM 版本，深度基于ESM 的模型进行定制